requirements.txt
streamlit
pdfplumber
googletrans==4.0.0-rc1
python-docx
fpdf 